// fync/src/github/types/github-actions.ts
type TGitHubWorkflowState =
	| "active"
	| "deleted"
	| "disabled_fork"
	| "disabled_inactivity"
	| "disabled_manually";

type TGitHubWorkflow = TBaseEntity & {
	name: string;
	path: string;
	state: TGitHubWorkflowState;
	badge_url: string;
	html_url: string;
	url: string;
	node_id: string;
};
// ...
type TGitHubWorkflowJob = TBaseEntity & {
	run_id: number;
	run_url: string;
	run_attempt: number;
	node_id: string;
	head_sha: string;
	html_url: string;
	status: "queued" | "in_progress" | "completed";
	conclusion:
		| "success"
		| "failure"
		| "neutral"
		| "cancelled"
		| "skipped"
		| "timed_out"
		| "action_required"
		| null;
	started_at: string;
	completed_at: string | null;
	name: string;
	steps: readonly {
		name: string;
		status: "queued" | "in_progress" | "completed";
		conclusion:
			| "success"
			| "failure"
			| "neutral"
			| "cancelled"
			| "skipped"
			| "timed_out"
			| "action_required"
			| null;
		number: number;
		started_at: string | null;
		completed_at: string | null;
	}[];
	check_run_url: string;
	labels: string[];
	runner_id: number | null;
	runner_name: string | null;
	runner_group_id: number | null;
	runner_group_name: string | null;
	workflow_name: string | null;
	head_branch: string | null;
};

type TGitHubActionSecret = {
	name: string;
	created_at: string;
	updated_at: string;
};

type TGitHubRateLimit = {
	limit: number;
	remaining: number;
	reset: number;
	used: number;
	resource: string;
};

type TGitHubRateLimitResponse = {
	rate: TGitHubRateLimit;
	resources: {
		core: TGitHubRateLimit;
		search: TGitHubRateLimit;
		graphql: TGitHubRateLimit;
		integration_manifest: TGitHubRateLimit;
		source_import: TGitHubRateLimit;
		code_scanning_upload: TGitHubRateLimit;
		actions_runner_registration: TGitHubRateLimit;
		scim: TGitHubRateLimit;
		dependency_snapshots: TGitHubRateLimit;
		audit_log: TGitHubRateLimit;
		code_search: TGitHubRateLimit;
	};
};